create get us eoy banner function
